Moved Error and SBSError to a separate module
